revert PacketForwarder split in favour of a single task that clears both channels on wake
